// src/app/store/company/types.ts
export enum ActionTypes {
    REQUEST_ADD_COMPANY = '@company/REQUEST_ADD_COMPANY',
    RECEIVE_ADDED_COMPANY = '@company/RECEIVE_ADDED_COMPANY',
    REQUEST_GET_COMPANY = '@company/REQUEST_GET_COMPANY',
    RECEIVE_GET_COMPANY = '@company/RECEIVE_GET_COMPANY',
    REQUEST_UPDATE_COMPANY = '@company/REQUEST_UPDATE_COMPANY',
    RECEIVE_UPDATED_COMPANY = '@company/RECEIVE_UPDATED_COMPANY',
    REQUEST_DELETE_COMPANY = '@company/REQUEST_DELETE_COMPANY',
    RECEIVE_DELETED_COMPANY = '@company/RECEIVE_DELETED_COMPANY',
}

export type Company = {
    name: string
}